:truck: Rename: Talk -> Opinion DB명 변경
