+++ src/main/java/com/ticle/server/opinion/domain/Heart.java
@@ -1,0 +1,34 @@
+package com.ticle.server.opinion.domain;
+
+import com.ticle.server.user.domain.User;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Table(name = "heart")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class Heart {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "heart_id")
+    private Long heartId;
+
+    @JoinColumn(name = "user_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    private User user;
+
+    @JoinColumn(name = "comment_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Comment comment;
+
+    @Builder
+    public Heart(User user, Comment comment) {
+        this.user = user;
+        this.comment = comment;
+    }
+}
