+++ src/main/java/com/ticle/server/opinion/domain/type/Order.java
@@ -1,0 +1,16 @@
+package com.ticle.server.opinion.domain.type;
+
+import org.springframework.data.domain.Sort;
+
+public enum Order {
+    TIME,
+    HEART,
+    ;
+
+    public static Sort getOrder(Order orderBy) {
+        return switch (orderBy) {
+            case TIME -> Sort.by(Sort.Order.desc("createdDate"));
+            case HEART -> Sort.by(Sort.Order.desc("heartCount"));
+        };
+    }
+}
