--- src/main/java/com/ticle/server/talk/dto/response/CommentShortResponse.java
@@ -1,17 +1,0 @@
-package com.ticle.server.talk.dto.response;
-
-import com.ticle.server.talk.domain.Comment;
-import lombok.Builder;
-
-@Builder
-public record CommentShortResponse(
-        String nickname,
-        String content
-) {
-    public static CommentShortResponse from(Comment comment) {
-        return CommentShortResponse.builder()
-                .nickname(comment.getUser().getNickName())
-                .content(comment.getContent())
-                .build();
-    }
-}
