--- src/main/java/com/ticle/server/talk/exception/CommentNotFoundException.java
@@ -1,12 +1,0 @@
-package com.ticle.server.talk.exception;
-
-import com.ticle.server.global.exception.errorcode.ErrorCode;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class CommentNotFoundException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-}
