+++ src/main/java/com/ticle/server/user/domain/User.java
@@ -2,7 +2,7 @@ package com.ticle.server.user.domain;
 
 import com.ticle.server.memo.domain.Memo;
 import com.ticle.server.scrapped.domain.Scrapped;
-import com.ticle.server.talk.domain.Talk;
+import com.ticle.server.opinion.domain.Opinion;
 import com.ticle.server.user.domain.type.Category;
 import jakarta.persistence.*;
 
@@ -14,7 +14,6 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @Table(name = "users")
@@ -49,7 +48,7 @@ public class User {
     private boolean agreeTerms;
 
     @OneToMany(mappedBy = "user")
-    private List<Talk> talks;
+    private List<Opinion> talks;
 
     @OneToMany(mappedBy = "user")
     private List<Memo> memos;
